Initial commit
This commit lays out a first implementation for a Sudoku solver.

In its current form, it seems to be able to solve Easy and Medium
NYT puzzles.

It fails on the hard puzzle because the implementation of
`solveNakedGroups` is incorrect: it assumes that the constituents
of such groups have identical possibilities, which is not the case
for groups larger than 2.

Committing here though before attacking that problem.