Implement solveNakedGroups
This diff implements a solution for finding naked groups.

It starts by iterating over the groups within the board.

For each group, it enumerates all the combinations of
unfilled cells.

For each combination, it calculates the union of possible values
that may be used to fill those cells.

If the number of possible values is equal to the number of cells
in the combination, then we have found what we're looking for:
a naked group. We then iterate over the other cells in the original
group and prohibit all the values found in group.

This now solves our three puzzles, and, further allows to delete
`solveNakedSingles`, which is a subset of the `findNakedGroups`.