Improve logging output
I added an indent "depth" field to the main methods that mutate
the board, `SetValue` and `ProhibitValue`, and increment the indent
each time one of these methods calls the other. This allows us to see
more clearly how effects cascade: when you set a value in the board, you
prohibit possibilities in other positions, and when you prohibit a
possibility, you may end up with only a single remaining possibility and
thus can set a value.

This lays bare the essence of how the solver works: there are a few ways
that you can find places to start, either by setting a value or
prohibiting a value. Eventually, the board gets to a state where one of
these starts leads to a volley between setting and prohibiting that
completes the whole board.