Remove intersections map
The intersections map is an optimization: it allows us to calculate
only once which rows and columns intersect each block and store
that result for later repeated use.

But it relies on pointers, which we're aiming to eliminate. So instead
we dump the intersections map but keep our blocks separate from our
rows and cols.