Implement board generator
This commits implements a board generator, which alternately
fills in a random cell and then attempts to solve the puzzle.

The algorithm is recursive to support back-tracking: if at
any point it picks a random cell value that leads to a conflict,
it backs up and tries a different value. It stops when it finds
a complete board.