Allow SetValue() to succeed...
... when the cell is already set to the given value.

There is a somewhat subtle relationship between the behavior
of SetValue() and CanTake(). We want to make CanTake() more strict,
i.e., if the value is already set, Cell.CanTake() should always
return false. But if someone goes ahead and tries to set a value
anyway, it's a no-harm/no-foul situation provided the value being
set matches what's already there.

Having CanTake() be more conservative in this way allows the solver
to skip some paths that it would otherwise explore. Having SetValue()
be more permissive allows to continue solving puzzles that are
over-determined, e.g., puzzles that are already solved before we
finish setting all the initial values.